Modernize archiver-api: split mgmt.rs into modules, add service traits
Split the monolithic mgmt.rs (1197 lines) into organized submodules:
- dto/mgmt.rs for DTOs and conversion functions
- handlers/mgmt/{pv_query, pv_control, reports, config_io, system}.rs
- handlers/health.rs for health and metrics endpoints
- state.rs for AppState and auth/metrics middleware

Introduce service trait abstractions (PvRepository, ArchiverControl,
ClusterRouter) with adapter implementations, enabling trait-object-based
dependency injection in AppState. Update all handlers, tests, and main.rs.

Fix all clippy warnings across the workspace.

+++ crates/archiver-core/src/retrieval/query.rs
@@ -18,7 +18,7 @@ pub async fn query_data(
         let desc = EventStreamDesc {
             pv_name: pv.to_string(),
             db_type: crate::types::ArchDbType::ScalarDouble,
-            year: chrono::Utc::now().year() as i32,
+            year: chrono::Utc::now().year(),
             element_count: None,
             headers: Vec::new(),
         };
